Fixing bugs of the pipe.
It was possible to open the pipe by catching a falling ball
inside the sensor, then attaching it anywhere, because in
OGContactListener::BeginContact() and OGContactListener::EndContact()
the same method was called and OGSensor::BeginContact() runs after
attaching the ball. The new member in OGUserData won't let sensors
balls_ decrease when doing the same (OGSensor::EndContact() is also
called after attaching) and avoids the pipe to remain open if a ball
enters the sensor attached but leaves unattached.
Setting the isTouching false in OGExitSensor::_EndContact() avoids
the force to keep applied if the ball was pulled away from the pipe.